% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table_align.R
\name{check_table_align}
\alias{check_table_align}
\title{check_table_align}
\usage{
check_table_align()
}
\description{
checks for date/time misalignment for tows between ILTS_SENSORS and ILTS_TEMPERATURE tables.

outputs to working directory a table showing number of TEMPERATURE rows found within SENSOR time range for each tow.

also outputs a filtered table only showing tows with 0 matching rows (no alignment).
}
\author{
Geraint Element
}
