% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilts.operations.R
\name{esonar2df}
\alias{esonar2df}
\title{esonar2df}
\usage{
esonar2df(esonar = NULL)
}
\arguments{
\item{esonar}{The esonar data to convert}
}
\value{
dataframe
}
\description{
Change esonar data format into useable dataframe
}
