% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilts.operations.R
\name{format.lol}
\alias{format.lol}
\title{format.lol}
\usage{
\method{format}{lol}(x = NULL)
}
\arguments{
\item{x}{latitude ot longitude values to convert}

\item{lol}{Specify if you are converting latitude or longitude}
}
\value{
list of correct latitude or longitude
}
\description{
Take latitude or longitude character vector and convert to decimal degrees.
}
