% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilts.operations.R
\name{get.oracle.table}
\alias{get.oracle.table}
\title{get.acoustic.releases}
\usage{
get.oracle.table(
  tn = "",
  server = pkg.env$oracle.server,
  user = pkg.env$oracle.user,
  password = pkg.env$oracle.password
)
}
\arguments{
\item{tn}{tablename to get}

\item{oracle.user}{your oracle username}

\item{oracle.password}{your oracle password}
}
\value{
dataframe
}
\description{
Get data from Oracle Database table
}
