% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterMarport.r
\name{LobsterMarport}
\alias{LobsterMarport}
\title{LobsterMarport: Reads data from non-tabular Marport reports and converts this into two separate tables containing GPS and sensor data}
\usage{
LobsterMarport(file, station = NULL)
}
\arguments{
\item{file}{is the marport file obtained during the 2015 ILTS survey which contains the net sensor information as well as the gps data feeds}
}
\value{
returns a list containing two elements: 1) the trawl sensor information 2) the gps track of the file
}
\description{
LobsterMarport: Reads data from non-tabular Marport reports and converts this into two separate tables containing GPS and sensor data
}
\author{
Adam Cook
}
