% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marport2esonar.R
\name{marport2esonar}
\alias{marport2esonar}
\title{this function converts the marport data tables produced by ILTSMarportSummary.r into Esonar format}
\usage{
marport2esonar(
  fpath = file.path("C:", "bio.data", "bio.lobster", "data", "survey"),
  year = year
)
}
\arguments{
\item{fpath}{is the path to where marport text data files are stored}

\item{year}{specifies the marport survey year for which R tables will be converted}
}
\description{
this function converts the marport data tables produced by ILTSMarportSummary.r into Esonar format
}
\author{
Geraint Element
}
