% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marport2esonar2021.R
\name{marport2esonar2021}
\alias{marport2esonar2021}
\title{marport2esonar2021}
\usage{
marport2esonar2021(
  data_directory = file.path(bio.datadirectory, "bio.lobster", "data", "survey",
    "Marport 2021"),
  output_directory = file.path(data_directory, "Marport converted"),
  correct.col.names = F,
  heading = "magnetic"
)
}
\arguments{
\item{data_directory}{is file path to where unconverted Marport files are stored.}

\item{output_directory}{is file path to create directory where converted data will be stored.}

\item{correct.col.names}{June-July 2021 Marport data had incorrect column headers, if this problem is not resolved in later outputs, use correct.col.names=T (function will override to T if June-July 2021 data is being run)}

\item{heading}{chooses whether to use "true" or "magnetic" heading. Default = "magnetic"}
}
\description{
for converting 2021 marport data to esonar format.
}
\author{
Geraint Element
}
