% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marport2esonar2023Fall.R
\name{marport2esonar2023Fall}
\alias{marport2esonar2023Fall}
\title{marport2esonar2023Fall}
\usage{
marport2esonar2023Fall(
  extract_data_directory =
    "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/ILTS Data Backup/2023/Leg 2/Marport/Marport Extract 2023",
  main_data_directory =
    "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/ILTS Data Backup/2023/Leg 2/Marport",
  output_directory = file.path(main_data_directory, "Marport converted"),
  correct.col.names = T,
  heading = "magnetic"
)
}
\arguments{
\item{extract_data_directory}{is file path where extracted NMEA files with Headline:OPN data are stored. THESE MUST BE IN THE SAME ORDER IN THE FOLDER AS THEIR CORRESPONDING MARPORT FILES!}

\item{main_data_directory}{is file path to where main unconverted Marport files are stored.THESE MUST BE IN THE SAME ORDER IN THE FOLDER AS THEIR CORRESPONDING NMEA FILES!}

\item{output_directory}{is file path to create directory where converted data will be stored.}

\item{correct.col.names}{2021 Marport data had incorrect column headers, if this problem is not resolved in later outputs, use correct.col.names=T to correct column names (function will override to T if 2021 data is being run)}

\item{heading}{chooses whether to use "true" or "magnetic" heading. Default = "magnetic"}
}
\description{
An update from the older function for converting marport data to esonar format.Includes merging of extracted NMEA files for Headline:OPN data.MAIN DATA (MARPORT) AND CORRESPONDING EXTRACTED DATA (NMEA) FILES MUST BE STORED IN SEPARATE FOLDERS AND IN THE SAME ORDER TO USE THIS FUNCTION!
}
\author{
Geraint Element
}
