% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmind2esonar.R
\name{netmind2esonar}
\alias{netmind2esonar}
\title{This function converts survey tables recorded in netmind format to Esonar format}
\usage{
netmind2esonar(
 
    main_directory = "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/Data & Code Files/ILTS Data/netmind_to_esonar"
)
}
\arguments{
\item{main_directory}{specifies path to main folder where netmind text tables are stored and where converted csv files will be outputed. Default ="R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/Data & Code Files/ILTS Data/netmind_to_esonar"). Directory must contain a folder called data containing the txt files and a folder called output.}
}
\description{
This function converts survey tables recorded in netmind format to Esonar format
}
