% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartfill.R
\name{smartfill}
\alias{smartfill}
\title{smartfill}
\usage{
smartfill(variable = "", reftime = "", timeflex = 4, rowflex = 30)
}
\arguments{
\item{variable}{The variable with NAs that the user wishes to fill with nearby values.}

\item{reftime}{specifies the time variable used to make sure that cells are not filled with values from across large time gaps}

\item{timeflex}{specifies (in seconds) how far away in time a nearby row can be to use its value.}

\item{rowflex}{species number of rows the function will search to fill an NA before it gives up (default=30)}
}
\description{
autofills NA cells of a specified variable in a time series data set with nearby (in time) non-NA values.
}
\author{
Geraint Element
}
