% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilts.operations.R
\name{click_touch}
\alias{click_touch}
\title{click_touch}
\usage{
click_touch(
  update = FALSE,
  user = "",
  years = "",
  skiptows = NULL,
  select.tows = NULL,
  dummy.depth = NULL,
  divert.messages = FALSE
)
}
\arguments{
\item{update}{TRUE: Add new startion and overwite any previous. FALSE: Skip previously completed set.nos.}

\item{user}{Unique user. This is used to keep user files seperate}

\item{years}{single or vector of years to process}

\item{skiptows}{skip specific tows specified by TRIP_ID:SET (example: skiptows = '100054289:51')}

\item{no.depth}{if real depth data is insufficient but still want to check headline distance and wingspread, run specified tows with dummy depth (example: no.depth = '100054289:51')}
}
\value{
list of lists. Format (top to bottom) year-set-data
}
\description{
Get all data, format and merge.
}
