% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilts.operations.R
\name{ilts.format.merge}
\alias{ilts.format.merge}
\title{ilts.format.merge}
\usage{
ilts.format.merge(update = TRUE, user = "", years = "")
}
\arguments{
\item{update}{TRUE: Add new startion and overwite any previous. FALSE: Skip previously completed stations.}

\item{user}{Unique user. This is used to keep user files seperate}

\item{years}{single or vector of years to process}
}
\value{
list of lists. Format (top to bottom) year-set-data
}
\description{
Get all data, format and merge.
}
